chatbot/requirements.txt
"Die folgenden Bibliotheken müssen mit 'pip install ...' 
der Umgebung hinzugefügt werden, damit der Chatbot funktioniert"  

langchain-core
langchain-community
langchain_ollama
langchain_chroma
python-dotenv
pyppeteer
pypdf
uuid
playwright 

"Die folgenden Modelle müssen mit 'olllama pull ...'"  

 mxbai-embed-large
 llama3.2